pyproject.toml
[project]
name = "digital-gym-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gspread>=6.2.1",
    "loguru>=0.7.3",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
